AST: finalised transpilation and handling of operators.
Operators will be transpiled into one of the three variants:

1). Named: Operators which don't exhibit any properties like side-effects, which
only logical 'and' & 'or' operators do. These function names undergo a simple
transpilation process where they are converted into the appropriate function calls.

2). LazyNamed: Operators which do have side-effects and need additional handling with the
rhs. Essentially, the rhs will be transpiled into a lambda call like '() => rhs' which
then enables the implementation of the logical short-circuiting operators to actually
exhibit the desired behaviour. Everything else is the same as the 'Named' transpillation.

3). Compound: Operators which need to be transpilled into either multiple function calls
or into a match case that can compare some kind of resultant Enum variant. Currently,
the only Compound variants are ones that involve the 'ord' trait. There are four
operators that use the 'ord' trait: '>=', '<=', '<', and '>'. These will be transpiled
into a call to 'ord' and then a match case that compares the result of 'ord()' to
the 'Ord' enum which represents ordering of two items. The match case is used to convert
the enum variant into a 'boolean' as the ordering operators exhibit this behaviour.

Additionally, any of the variants might be 'assiging', or in other words an operator
that will overwrite the value of 'lhs' by applying some intermediate operation. If this
is the case, we convert the 'lhs' into a reference expression so that we can perform
the intermediate operation in one go.

Overall, this is a simplification over the Haskell implementation and expands the capabilities
of operator overloading.