AST: Use identifier caching to prevent overhead from string operations.
Before, each identifier in the AST (like in Name and AccessName) was a
string (Cow, granted), which wasn't ideal. This commit makes the
following change:

Identifiers are kept track of using a static map structure that assigns
IDs to each string identifier. Then, these IDs can be used in the AST
nodes instead of strings, to represent identifiers. This will also be
useful during program analysis, when checking for identifier equality.

Furthermore, this commit removes type_id from the AstNode struct and
moves it to Expression, which is the only node that actually requires
it.