Integrated logging via the '-d/--debug' cmdline options.
The library 'pretty-env-logger' doesn't support variable logging levels during runtime. To workaround
this limitation, we now implement a `log::Log` which does allow for this behaviour. The logger is initialised
in the same way as before, but if 'debug' is specified via cmdline args, the level is set to `debug`.