WIP: Split some code between `resolve.rs` and `module.rs` in hash-ast.
This is just the beginning of an effort to simplify and improve the
way the parsing happens. We want to move away from a channel based system
and just use threads with a queue of jobs.

Moving code out of parse.rs that wasn't relevant to the actual parsing was
also done. Everything that is related to representing `Modules` was moved
into `module.rs`, and everything to do with resolving modules was moved into
`resolve.rs`.