Tokeniser: treat opening delimeters ('('/'{'/'[') as individual token trees.
With the new `Tree(..)` variant, the tokeniser creates sub trees within the structure
to record the beginning of such trees, and will advance tokens until it reaches the closing
corresponding delimiter or EOF. If it reaches EOF before a closing delimiter, this is considered
to be a syntax error and is reported by the tokeniser.

The reason why the tokens are now in the form of trees is due to the fact of parallelisation efforts.
Now that we have the trees, we can give threads/jobs these trees to process on their own time into
corresponding AST. This means that the transformation process is now able to be far more parellisable
than by each file, it will be able to be on a block by block basis (this will greatly improve performance).

Additionally, this change also mitigates delimiter depth syntax checking issues, because this naturally
counts the depth in the sense of recursion. We can avoid keeping track of this information later on in the
AST gen.