AST: Do not copy AST token trees when parsing.
Beforehand, the AstGen struct owned its token stream in the form of a
Row. This meant that every time a new token tree was encountered, it had
to be cloned into the child AstGen struct. To reduce cloning, AstGen now
stores a slice of the token stream, so no copying is required at any
point during parsing.