AST: operator re-assigning and transformation of compound operators is now supported.
As previously implemented in the PEST parser backend, we have some logic to
transform operators into function calls. Depending on the actual operator and
if it is re-assiging (in the sense that it's using the lhs as an argument, e.g: a += b),
we need to transform them in specific ways.

This logic is now implemented in the self hosted parser and exhibits the same
behaviour as the PEST parser.