Parsing: made ParsingError location optional.
There is a case (and maybe others) where the location of the parsing error is unavailable due to module importing or some other edge case. In this situation, the reporting for parsing error's shouldn't rely on the default `Location` value and attempt to read a file that doesn't exist in the module builder (#61).