pest: fix improper error message formatting when converting into a ParseError.
When the pest parsing backend attempted to print an error, it would just call `to_string` on the
`pest::Error::ParsingError` which used it's own printing mechanism and would print a weirdly
formatted message. This patch takes printing the pest error into our own hands and makes it
more conformant to how we expect to print error messages instead of relying on pest as we did
before.

This also prompted a change to unify some display printing logic between multiple crates
in the compiler. Thus, a new module in `hash_utils` called `printing` is added to contain
useful methods and implementations for pretty printing some types. Currently, the module
contains an implementation to print a `SliceDisplay` which is a wrapper around a slice
in order to print in a English-like way. Given the following a slice:

>>> ["tree", "apple", "worm"]

It would display it as the following string:

>>> either a tree, apple, or worm.