Add struct definition parsing and fix memory bug with Brick
Brick::disown* family is not safe because it takes self by value, which
means that its destructor will run at the end of the scope. This
invalidates the value and causes unexpected behaviour. It is preferred
to use wall.alloc_value instead, which provides the same functionality
without the direct usage of unsafe in hash-alloc.