refactor: move `Location`, `SourceLocation` and `ModuleIdx` into hash_source.
This change is necessary in order to sort out a dependency hell between
`hash_reporting` and any modules like `hash_ast` that use it to
construct a report and then print it. It wasn't particularly ergonomic
to use the `From<SomeError> for Report` because then it means that
`hash_reporting` depends on x-many crates to implement the
logic for converting an error into a report.

The alternative was to move logic for converting into their origin
crates. This meant that a method is present on each error called
`create_report` which converts the error into a `Report` (since
Rust doesn't allow for implementation of `From<..> for Report` in
foreign crates).

This caused an issue with cyclical dependencies between `hash-ast` and
`hash-reporting` since reporting depends on `Modules`, `SourceLocation`
and `Location` to get a particular span from a source. The solution
is to move all of this logic and de-couple it from any kind of
`ast` dependant stuff.