fix: split messaging of setting source and node for a module
This brings a fix for the case where the initial module or
any other module isn't added into the module map before the
reporter needs to the module for it to print the error.
This changes the mechanism in to add the source before
even lexing so that it's always available to print the error.