fix #113: Malformed error report.
Added new variant for `AstGenError` which occurs when the
parser expects an identifier, but receives something unexpected.
Previously, the parser would report that it expects an expression
which isn't particularly accurate.

Additionally, the location of the error was slightly incorrect and thus
has been adjusted to point at the next token instead of the current
one as in that context the parser was peeking.