Simplify compiler pipeline and properly implement enums.
- Always return the printed type from `run_interactive` whether requesting
or not. This simplifies the pipeline and removes the redundant functions.

- Implement enums properly: added a new `SymbolType` to represent an
enum variant variable (currently added into parent scope). Then when
making a new variant, first you make a substitution with the enums
type arguments to its variant data members. After the application,
the instantiated type is returned. For the case where the variant
has no data members, the type of the enum is simply returned. If
the variant however does have members, a function type is created
that takes the data members as arguments and returns the type of the
enum.