typechecker: Implement tuple types.
Modify the AST to have a special variant for a tuple type. It was difficult
to differentiate between a "Named" type and a tuple type, and so we
make the clear distinction now between this.

This patch also fixes some inconsistencies with the way displaying the
AST and tuple types.