AST: use special variant for function types.
Similarly to tuple types, the current way of handling function types
was to check if an identifier was being used to denote the function
as 'Function' which was very hacky. This patch removes this behaviour
and adds a new variant to the `Type` variant called `Fn`.

This makes it considerably easier to work with in the typechecker
and the visualiser.