typechecker: Implement match block type checking.
Implemented typechecking of match blocks with a crude implementation of
`visit_match_block` and `visit_match_case`. The current implementation
verifies that all patterns must match the subject type and that all
expressions must unify to one type which are the current rules of match
blocks.

Moving onwards from this, we need to figure out how we deal with `if`
statements that lack an else case and receive one automatically. This
is problematic because the auto-generated `else` branch will always
have a type `void`. This is fine in a lot of cases, but I think there
a few edge cases that should be figured out.

Additionally, there is still `visit_struct_pattern`, `visit_enum_pattern`
and `visit_namespace` pattern left to implement.