fix: Report error when missing comma after '=' in struct literals.
When struct literals were missing a comma somewhere in the middle of
the brace block, the parser would abort parsing any fields after it
because it considered that there cannot be anything after the
comma. However, this wasn't the correct behaviour as there might
be just a missing comma afterwards, and it will just fail to
parse everything else in the block. This patch fixes this
behaviour and reports the error.

This also improves the way token kinds are being reported in the
parser reporter logic to make more sense (in terms of English).