Parser: `()` is now treated as an empty tuple (fixes #147)
Previously, the parser did not expect empty parenthesees in the form
of `()`. This behaviour was inherited from the Haskell implementation
since at the time we wanted to eliminate the ambiguity between
this kind of expression and a tuple at the syntax level. However,
this doesn't seem relevant in this case because if the parenthesees
have no inner expression members. For this reason, expressions that
are in the form of `()` are now labelled as empty tuples.