Parser: Change syntax of fn return types as per #151
This patch addresses the issue specified in #151. Annotated
function return types are now denoted using a '->' for both
in function literals and within function types.

Before, there was a different syntax for specifying the return
type of a function within a type (using the '=>' fat-arrow) and
within function literals (':' colon). This was inconsistent and
sometimes hard to read. This change syntax should clear up
the ambiguity of annotating return types within functions.