report: fix issue with operator locations being under reported.
This patch fixes an issue with operator locations such as '<'
not being properly recorded due to the function `transform_compound_ord_fn`
not taking the operator location into consideration. This patch makes
the function take in the operator kind and location.

Additionally, in `hash-typecheck`, the reporting for the symbol
resolution was referencing the whole `VariableExpr` location
instead of just the location of the subject of the expression.

fixes #159