Typechecking: Fix boolean patterns
The fact that "true" and "false" are variables and not enum variants
means that they cannot be used within a pattern, and instead they are
assumed to be pattern bindings, which overwrites them with whatever the
subject of the match case is. This results in unexpected behaviour such
as:

```rs
match 1 {
    true => {
        // Now `true` is of type `i32`
    }
}
```

The actual solution would be to make booleans enums, which should
automatically fix this problem. However, this is a quick fix to get
match statements (and following from that, if and while) to work.