pipeline: Allow for compiler to have modes
This patch expands the `hash-pipeline` to have modes that are
derived from the provided command-line arguments. By doing this,
the compiler now supports `ast-gen` mode again which will only
run the pipeline up to the typechecking stage.

With `ast-gen` mode, this also means that when the compiler runs
with the `--debug` flag, it will print the generated AST for
each module source that has been parsed.