Parser: drop `let`s from declarations and use `:=` syntax
This patch is to switch to the new declaration syntax which avoids
using the `let` prefix and then using a pattern and `=` to optionally
assign a value.

The new syntax uses `a := ...;`, more specifically the `:=` denotes
that this is a new declaration rather than a assignment. The programmer
can specify the type if required between the `:` and `=` as it
was previously possible with the let syntax. For example:

>>> let x: u32 = compute_something_which_returns_u32();

Now becomes:

>>> x: u32 = compute_something_which_returns_u32();

Or even:

>>> x := compute_something_which_returns_u32();

There are some obvious complications with this method as it now
becomes difficult to differentiate between expressions and declarations
(The error reporting is still not quite there even now).

The general parser steps to differentiate between declarations and
expressions is:

 1. Attempt to parse a declaration pattern which is a subset of all
    the possible patterns that can be parsed.

 2. If this fails, then an attempt is made to parse an expression and then
    a re-assigning operator (or even just an assignment operator).

 3. If this is successful, then we continue assuming that this is a
    declaration. If however, the next atom is not a `where` or `:` then
    we backtrack entirely and then go to step 2.