Restructure documentation structure
This commit changes the structure of the Hash language documentation to
be more in-line with the latest syntactical changes to the language.
The new section structure can be found in `docs/src/SUMMARY.md`.
Formal CFG grammar is given for each language construct which should
hopefully make it easier to implement parsing and have fewer ambiguities
in the language.

For now, "variables", "primitives" and part of the "functions" sections
have been rewritten. Subsequent commits will tackle the rest of the
documentation.