Remove unnecessary ":=" variant in type function arguments
It is not clear how to infer the type of a type, because it can be any
level of refinement.
For example, if `T` implements `Copy` and `Eq`, the following are true:
```
T: T
T: Copy
T: Eq
T: Copy ~ Eq
T: Type
```

Given this, which bound should be inferred? The answer is that we
shouldn't try to infer a bound, and it should be `Type` by default
unless someone specifies otherwise (with an explicit `: MyBound`).