Parser: Add support for `type` expressions
This patch adds support for parsing type expressions. They follow
a simple syntax of `type <type>`.

This patch fixes a bug in the parser which prevented a subset of
valid expressions not being parsed. The bug was that type function
definitions always expected the parameters to have type bound
annotations. Since this isn't true, this was updated to reflect
this fact.