AST: Add mutable visitor and walker
The AST lowering pass requires a visitor and walker methods that allows for a mutable reference within the visitor so that the pass can transform the specific nodes and in place.

Side note:

This is a dogshit commit, this does not indicate the abilities of the maintainers. This is Rust's fault and not mine. cheers