Somewhat restructure the organisation of types
This commit introduces the concept of kinds and values, which are
present for each binding that exists in the program (i.e. each member of
a scope). This is because we need to be able to differentiate between a
type and a value of a given binding. Some docs are added to explain the
change.