Slightly restructure how scopes are stored and start `SymbolResolver`
Scopes are now stored in an actual store and are indexed by IDs. This
makes it easier to handle scopes without rust's reference aliasing rules
getting in the way. Also `Members` has been renamed to `Scope`, and the
old scope stuff has been mostly removed. The beginnings of a
`SymbolResolver` have been implemented in the `ops` module.