Refactor the usage of both `Ty` and `Value`, and use only `Term`
While implementing the unifier it became clear that the distinction
between Ty and Value is entirely arbitrary, since code was duplicated
for both, in terms of substitutions, simplifications, unifications,
etc.

It is better to have a single enum `Term` which can be used both in type
position and value position. It is split up into levels, which determine
the different levels of the Hash type system (consult initial TC PR for
more info, as well as the docs), as well as some general terms whose
level depends on their contents (access, type functions, application,
unresolved, etc).

With this done, hopefully now unification will become much simpler/DRY
and easier to reason about.