parser: introduce `set` and `map` syntax
This patch adds the `set` and `map` syntax to denote literals rather than dealing with the `set/map/block` ambiguity.

This also cleans up alot of the token tree parsing with introducing the `parse_delim_tree` function that is general purpose for parsing the next token and producing a `AstGen` generator from it.