pipeline: continue running stages when diagnostics are only warnings
This patch re-works the pipeline in order to cope with the fact
that some diagnostics might be warnings, and it is ok to
continue compiling. This also generalises how the pipeline
deals with diagnostics from all the generated stages.

The pipeline now collects errors within `compiler_state` and
only prints them when the pipeline terminates. It also reports
how many warnings and errors were generated when running a particular
job.