ast: introduce `MethodCall` expr variant for infix calls
This patch introduces a new expression variant `MethodCall` in order
to represent infix calls that are parsed by the parser. Previously,
the parser would automatically de-sugar these infix-like calls
into normal calls. Now that the typechecker can interact
with these constructs, it is easier if it received them as they
were parsed, not already de-sugared.

closes #372