parser+ast: Add support for `union` types
This patch adds support within the AST and parser for
union type notation. The patch introduces the notion
of type operators which are equivalent to binary operators
but for types. Currently, `TyOp` has two variants which
are `Union` and `Merge`, where union operator has a
higher precedence then the merge operator.

closes #362