parser: Implement multi-error handling!
This patch adds some support for emitting multiple errors with
reasonable recovery capabilities. Most of the heavy work is
done by the lexer since it creates token trees that correspond
to a delimited stream of tokens. This means that we can create
a pattern within the parser to create a child `AstGen`, run
functions on it that might emit errors and then capture them
into our current `AstGen`.

The function `parse_separated_fn` does most of the heavy lifting
since it will capture errors rather than emitting them into the
passed in generator. It is now also able to do some basic error
recovery when it cannot parse the `item` or the `separator`.
Then, because `parse_separated_fn` produces a `AstNodes<T>`, we
don't necessarily care about the contents of the result since
we will emit the errors anyway. By applying this strategy we
can greatly enhance the robustness of the parser and at relatively
little cost.

There are plenty of other places where the parser could be made
even further robust by adding more error recovery logic, but for
now this will suffice.

Places that have been marked with `@@ErrorRecovery` are the
low hanging fruit.