tc: improve printing of integer literal terms
This patch improves how integer literals are displayed when
being printed within typechecking. Now the type of the integer
and the literal itself is separated with a `_`.

Additionally, integers that are `MIN` and MAX` values of a
particular integer range are printed as `MIN` and `MAX` for
better visual clarity when reporting exhaustiveness errors.