tc: implement `...` pattern erasure for tuple patterns
Tuple patterns that feature a `...` inner pattern are now erased
from the tuple. This means that any patterns that the `...` pattern
captures are converted into wildcards. If the spread pattern features
a bind like `...x`, `x` becomes a tuple type of all the captured
fields which can be accessed.

This can be better explained with a concrete example:

```
main := () => {
    t := ("Alex", 18, 'c', 183.1);

    (k, ...x) := t;         // x=(i32, char, f32)
    (...x, k) := t;         // x=(str, i32, char)
    (k, ...x, l) := t;      // x=(i32, char)
    (k, c, ...x, l) :=      // x=char
    (k, c, ...x, l, m) := t // x=()
};
```