tc: fix bug with `term_of_pat` overwriting existing term locations.
When calling `typer().get_term_of_pat()`, the location would be
unconditionally copied to the term of the pattern. However, this
is not the correct approach since for `Pat::Const`, the term already
exists and has it's own location. This was incorrect, and now the
location is only copied if the branch specifies that copying
the branch is needed.