tc: fix constructor fields not being added to scope and ordering of field mismatching
This patch fixes how constructor arguments and the associated struct
type were being matched when fields were being iterated in the
`PatMatcher`.

Additionally, this fixes a problem with fields that don't have
binding patterns not being added into the scope as members. This
should work since each specified field in the constructor pat
should be added to the scope unless it is explicitly ignored using
the `...` pattern.