tc: fix mis-fire of `NamedTupleCoercion` warning
The warning was mis-firing on tuple patterns that are only made
of a single `...` pattern, which meant all of the members are
un-named and thus the warning would fire if the subject type
is a named tuple.

Now, the fact of whether the tuple pattern is named or not is inherited
from the naming convention of the type in the situation where the
spread pattern is a single `...`.