Add unit nominal primitive
This commit adds `NominalDef::Unit`. This is meant to be to constant
enum variants what structs are to enum variants with data. Hopefully
that makes sense. It is equivalent in spirit to a struct with an empty
field list, or a simply a unit struct in rust `struct Foo;`.

In subsequent commits, this will be used to replace enums; enum variants
will become either structs or units, depending on whether they have data
members or not.

Closes #501.