source: introduce new `RowCol` and `RowColSpan` data structures
Previously, when the compiler needed to convert `SourceLocation`s
and `Span`s into their respective row and columns values for line
number information, it was being done as a second thought in the
reporting sub-system.

Now that there has been the need to access this logic from other crates,
it has been decided that this logic should live in `hash-source` and
be accessible by everyone who requires converting locations into
row and column values. This means that:

- computing `row` and `column` values logic now lives in `hash-source`.
- `RowColSpan` is a structure that represents a `Span` but with row
and column values.
- `RowCol` is a single unit data structure to group a `row` and
`column` value.