It is now clear that each stage cannot depend on some global "compiler state" since this would create crate dependency problems, and in general would make crates be un-clean.
The new approach allows for a `CompilerStage` to specify a trait that provides an interface for querying information from the `CompilerSession` which will be able to retrieve all of the information the stage needs. This way, all stages can remain "opaque" to `CompilerSession` and vice-versa.

This meant that a new crate called "hash-session" which contains the global compiler session, and all of the implementations for the stage specified context interfaces.

Additionally:
- moved `NodeMap` out of `hash_pipeline` and into `hash_ast`