ir: add `name` to `Body`
This patch adds a `name` to the body which is assigned during the
lowering discovery process in which the visitor will find
function definitions that need to be lowered and proceed
to do so. The current (ad-hoc) bind discovery method works on by extracting
all of the binds from a given declaration bind and pushing them
on a bind stack, which is then accessed to retrieve the
correct names.