Add the ability to assign `ScopeId`s to AST nodes
This commit adds the ability for AST nodes to have attached scopes, in the same
way that they currently can have attached `TermId`s or `PatId`s. This is
because scope information is required during code generation. The way to access
`ScopeId`s is through the `NodeInfoStore` struct on `GlobalStorage`, by
providing an `AstNodeId` as an input and retrieving a scope using the
`scope_id` method on the resultant `NodeInfoTarget`.

The following nodes have scopes attached to them:
- `MatchCase`
- `BodyBlock`
- `TraitDef`
- `TraitImpl`
- `ModBlock`
- `ImplBlock`
- `Modules`

Therefore, it is only valid to call `scope_id` on `NodeInfoTarget`s accessed
through an `AstNodeId` corresponding to the above AST nodes. (Otherwise there
will be a panic.)