ir: add `graphviz` output format, enabling IR visualisation
This commit implements a new printing backend for the IR which
will emit a graphviz dot file. This can be used to visualise the
IR in a variety of ways, including using the `dot` command line
tool to generate a PDF or PNG.

This also cleans up the implementation for the IR Body pretty printer
in order to share as much logic as possible between the pretty printer
and the graphviz printer.