core: introduce `LoweringSettings` and upgrade `clap` to `v4.0.0`
This commit introduces a new `LoweringSettings` struct that is used to
configure the lowering process. The user can now configure whether
the whole IR is dumped (and with what mode), and whether the
lowering process should use `checked` operations, amongst other
settings.

This patch also migrates the `clap` dependency to `v4.0.0`, which
cleaned up some of the argument handling code.