ir: add `BodyInfo` struct to group metadata information about `Body`s
This patch adjusts the way that metada is stored in `Body`s. Previously,
many of the fields in `Body` were used to store metadata about the body
itself. This patch introduces a new struct, `BodyInfo`, which is used to
store this metadata. This allows us to remove many of the fields from
`Body` and instead store them in `BodyInfo`. Additionally, `BodyInfo`
is meant to be built up as the IR body is being built to avoid having
to pre-compute everything before construction.

This meant that the functions that lower `hash-types` types are now
part of the builder and led to less code when lowering `hash-types`
into ir types, notably constantly passing in the `tcx` and the `storage`.