parser: don't overeagerly parse tuple literals when encountering `==`
This patch fixes an issue with tuple literal parsing where the parser
would attempt to parse a tuple literal value with a named field
being assigned for the following expression:

```
(x == 1)
```
The parser thinks that `x` is the name of the field, and the first
`=` is the assignment operator. This patch fixes this issue by checking
that the next token is not an `=` before parsing the tuple literal.