ir: introduce new constant variant that represents un-evaluated items.
When a body references an item from a constant scope, this value is
then recorded as the scope that it comes from, and the name of the
item. Later on, these values will become **evaluated** and are
essentially inlined into the place where they are used.

This is a step towards allowing constant items being used within
function bodies, however the way the reference is recorded will be
changed to a `DefId` or `DeclId` which refers to the origin declaration
that the item was defined from.