lower: introduce IR optimisation stage and implement a simplification pass
This commit introduces a new stage to the lowering pipeline which
is responsible for orchestrating optimisations on the IR. The first
implementation is not currently sophisticated and cannot be
parallelised, however it is a good starting point for future when
considering this problem.

The first pass implemented is a simplification pass which trims dead
blocks from the graph, and removes redundant jump chains. There are
many other possible and more sophisticated optimisations that
can be performed on the IR before code generation, however this
commit is a good starting point for future work.