lower: add initial scaffolding for `cleanup_locals` pass
This patch introduces a new pass called `cleanup_locals` which is
responsible for pruning the CFG from un-used locals. In principle, this
means that the locals are never assigned to, and or they never appear
within [RValue]s which means it is safe to remove them from the
[Body].

However, this pass is not yet implemented due to the IR Visitor not
being ready yet. This patch is just the scaffolding for the pass.