ast: simplify AST spread patterns
Spread patterns are now present as optional fields on AST Patterns
that allow for a spread pattern to occur. The pattern is also no
longer a standalone pattern member, but rather a member of a
particular compound pattern (namely `List`, `Tuple` and `Constructor`).

This change simplifies the AST, makes the spread pattern more robust
and reduces the amount checking that needs to occur at the untyped
semantic check stages.

Additionally, it is also very simple to identify if one these patterns
has a spread pattern because the spread pattern is now an optional
field on the pattern, and isn't somewhere in the middle of the pattern.

There are still some amendments needed to the type checking in
order to integrate this new change.