lower: massively improve performance of graph simplification pass.
Previously, this pass would continuously thrash the `predecessors`
cache on the CFG when updating basic blocks. This meant that
every time a block was updated, the whole graph predecessors
map would be re-computed. This means that it would be re-computed
per update, per block elimination, per pass. This is obviously very bad.

Now, the pass only computes predecessors once, and keeps an up-to-date
record as it eliminates dead blocks. Additionally, the pass now
tries to eliminate `goto` chains as it goes, which means it can
eliminate a lot more blocks in one pass, and thus needs to perform
fewer passes to scrub the entire graph.

We went down from a `550ms` to a measly `1ms`!