vm: remove interface impl for `CompilerStage`
Due to the new work with code generation, it is likely that the VM itself
does not implement the `CompilerStage` interface since it won't be
directly invoked by the compiler, but the actual `HashBackend` stage. The
backend stage decides whether it will need to invoke the VM for
some purpose, or if it can just generate code directly and emit
an executable.