utils: Add generic graph interface for various algorithm
Utility crate now hosts a generic graph interface that allows
for various algorithms and interfaces to be implemented for graphs.
Notably, the current interface provides a dominators algorithm and
some accompanying traits for traversing the graph.

It is likely that some code from `hash-lower::traversal` will be
moved into this crate too.