codegen: Begin writing `LocalKindAnalyser`
A `LocalKindAnalyser` is used to determine what kind of memory model
can be used to represent a `Local` when lowering. The kinds include:

```rust
enum LocalKind {
    /// Zero-sized type local, isn't constrained to any memory location
    Zst,

    /// Not ever referenced
    Unused,

    /// Singly-statically assigned, has only one use and assignment
    /// in all dominating blocks of the CFG.
    SSA(IrRef),

    /// Referenced/re-assigned multiple times within the CFG.
    Memory
}
```

The `LocalKindAnalyser` will make a pass on the IR body and compute
whether each local can be represented as a SSA value or if it needs
to be represented as a memory location.