codegen: implement slight optimisation/hack for codegening switch terminator edge case
In the event that a switch case has two targets, and the optional
one is unreachable, then the switch case uses an integer comparison
to select which basic block to jump to rather than a switch case.

This is useful for debug build speed-up on LLVM backend due to it
using different instructions (that are faster for codegen).

Consequently, the builder now has access to `CompilerSettings`.