layout: expand `LayoutStorage` to store layouts, and other information
This patch adds `LayoutStorage` to the general `CompilerSession`
struct which means that we can now store layouts in the session, and
access them across various stages of the compiler.

This is both needed for the code generation stage, but also for the
lowering stage in order to read and write the layouts of types. In
particular, the `#layout_of` directive will be able to query the
layout of a particular type, and thus pretty print it.