Separate typechecking from semantics, and improve diagnostics
The typechecking is now independent of the rest of the semantics pass,
and lives in a separate crate.

The diagnostics API has been changed slightly to `Diagnostics` and
`AccessToDiagnostics` traits. This is useful to access diagnostics methods
from a root object, without having to specifically refer to the
diagnostics implementor.