cli: write out custom logic for parsing commandline arguments
This introduces a more standard configuration argument system for specifying
options to the compiler in the form of `-C option-key=option-value`. This
means that configurations can be arbitrarily added to the commandline
options.

It was decided to move away from `clap` due to the fact that it does
provide enough flexibility to implement the desired configuration
argument system.

This patch removes clap from the compiler (which also shrinks the
release binary by ~450kb & improves compile times). This implements
the new argument parsing within the `hash-pipeline` crate. However,
this system isn't complete because there is no help menu system yet.

Additionally, this patch cleans up the error kinds within `hash-reporting`
which didn't really belong within the `errors` module, and they've been
split up amongst the `hash-pipeline` and `hash-interactive` crates.