backend: Add `Backend` trait
This commit begins the process of implementing an interface
between a code generation backend and the rest of the compiler
pipeline. Since it is possible for there to be multiple backends, it
is not known at compile time which backend will be used. Therefore,
the `Backend` trait is used to abstract away the details of the
backend.

The trait specifies that the the backend must be created
from a `BackendCtx` which should provide all relevant information
to the backend, and a `run()` method which initiates the process
of code generation. The `run()` method is responsible for configuring
the backend (if need be), and then generating code from the IR using
the common code generation interface defined in `hash-codegen`.

After this process is complete, the backend will update the `Workspace`
with the results of the code generation process. This is done by
adding information about what the code generation backend has generated
in the form of object files. After this, the linking stage can take
all of the object files and link them together.